+++ app/localdb/dbtables.py
@@ -1,7 +1,25 @@
 import uuid
 from typing import Optional
+from datetime import timezone, datetime
 
-from sqlmodel import Field, Relationship, SQLModel
+from sqlmodel import Column, SQLModel, Field, DateTime, Relationship, TypeDecorator
+
+
+class TZDateTime(TypeDecorator):
+    impl = DateTime
+    cache_ok = True
+
+    def process_bind_param(self, value, dialect):
+        if value is not None:
+            if not value.tzinfo or value.tzinfo.utcoffset(value) is None:
+                raise TypeError("tzinfo is required")
+            value = value.astimezone(timezone.utc).replace(tzinfo=None)
+        return value
+
+    def process_result_value(self, value, dialect):
+        if value is not None:
+            value = value.replace(tzinfo=timezone.utc)
+        return value
 
 
 # Self referential model (https://docs.sqlalchemy.org/en/latest/orm/self_referential.html)
@@ -32,12 +50,18 @@ class PasswordGroups(SQLModel, table=True):
 
 class PasswordEntry(SQLModel, table=True):
     entry_id: uuid.UUID = Field(primary_key=True, default_factory=uuid.uuid4)
-    entry_name: str = Field(min_length=1, nullable=False, index=True)
+    title: str = Field(nullable=False, index=True)
 
-    entry_username: str = Field(nullable=False)
-    entry_password: str = Field(nullable=False, min_length=1)  # keep this encrypted
+    username: str = Field(nullable=False)
+    password: str = Field(nullable=False)  # keep this encrypted
     
-    entry_url: str = Field()
+    url: str = Field(nullable=False)
+    notes: str = Field(nullable=False)
+
+    created_at: datetime = Field(
+        default_factory=lambda: datetime.now(timezone.utc),
+        sa_column=Column(TZDateTime)
+    )
 
     group_id: uuid.UUID = Field(foreign_key='passwordgroups.group_id', ondelete='CASCADE')
     group: PasswordGroups = Relationship(